Move misc serializers from `utils` into `misc` package
