+++ language-support/jobrunr-kotlin-21-support/src/main/kotlin/org/jobrunr/kotlin/serialization/storage/BackgroundJobServerStatusSerializer.kt
@@ -6,8 +6,8 @@ import kotlinx.serialization.descriptors.buildClassSerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.encoding.encodeStructure
-import org.jobrunr.kotlin.serialization.utils.DurationSerializer
-import org.jobrunr.kotlin.serialization.utils.InstantSerializer
+import org.jobrunr.kotlin.serialization.misc.DurationSerializer
+import org.jobrunr.kotlin.serialization.misc.InstantSerializer
 import org.jobrunr.storage.BackgroundJobServerStatus
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
