+++ language-support/jobrunr-kotlin-21-support/src/main/kotlin/org/jobrunr/kotlin/serialization/storage/JobStatsSerializer.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.encoding.encodeStructure
 import org.jobrunr.kotlin.serialization.utils.ClassDiscriminatedContextualSerializer
-import org.jobrunr.kotlin.serialization.utils.InstantSerializer
+import org.jobrunr.kotlin.serialization.misc.InstantSerializer
 import org.jobrunr.storage.JobStats
 
 object JobStatsSerializer : KSerializer<JobStats>, ClassDiscriminatedContextualSerializer.PolymorphicContinuationSerializer<JobStats> {
